etl/requirements.txt
# ETL Pipeline Dependencies
# Python 3.12+

# Data validation and schemas
pydantic>=2.0.0

# Supabase client
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0

